% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADLtired.R
\docType{data}
\name{ADLtired}
\alias{ADLtired}
\title{Physical Activities of Daily Living}
\format{
A data frame with 734 rows and 97 variables of which 19 are used.
The data was collected from 734 70-year old in the County of Copenhagen, Denmark.
The PADL scale consisted of a total of 16 items covering three different domains.
\describe{
Mobility function
\item{A:}{Are you able to walk indoors?}
\item{B:}{Are you able to walk out of doors in nice weather?}
\item{C:}{Are you able to walk out of doors in nice weather?}
\item{D:}{Are you able to manage stairs?}
\item{E:}{Are you able to get outdoors?}
\item{F:}{Are you able to get up from a chair or bed?}
Lower limb function
\item{G:}{Are you able to wash thelower part of the body?}
\item{H:}{Are you able to cut your toenails?}
\item{I:}{Are you able to go to the toilet yourself?}
\item{J:}{Are you able to dress the lower part of the body?}
\item{K:}{Are you able to take shoes/stockings on/off?}
Upper limb function
\item{L:}{Are you able to wash the upper part of the body?}
\item{M:}{Are you able to cut your fingernails?}
\item{N:}{Are you able to comb your hair?}
\item{O:}{Are you able to wash your hair?}
\item{P:}{Are You able to dress the upper part of the body?}
}
}
\source{
The \code{ADLtired} data were obtained from
Avlund, K., Schultz-Larsen, K., Kreiner, S. (1993) Construct validation and the Rasch model: Functional ability of healthy elderly people. \emph{Scand. J. Soc. Med.}, 21: 233-245.
}
\usage{
data(ADLtired)
}
\value{
ADLtired is an object of class digram.object.
}
\description{
Measure of functional ability of healthy elderly (Avlund et.al., 1993)
}
\keyword{datasets}
