% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{all.patterns}
\alias{all.patterns}
\title{Create all possible response patterns}
\usage{
\method{all}{patterns}(maxscores = c(), target = NULL)
}
\arguments{
\item{maxscores}{vector of max scores}

\item{target}{target total score}
}
\value{
Returns a matrix of all possible patterns. If target is set, only patterns with at total score of target is returned
}
\description{
Create all possible response patterns
}
\details{
This function quickly gets very resource intensive. On a 8 core, 16 GB computer, 25 items are too heavy.
}
\examples{
maxscores<-c(1,2,3,1,2,3)
all.patterns(maxscores,target=5)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
