% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.object.R
\name{as_tbl_graph.digram.object}
\alias{as_tbl_graph.digram.object}
\title{Create a tidygraph object from a digram.object}
\usage{
as_tbl_graph(do)
}
\arguments{
\item{do}{A digram.object.}

\item{LD}{A data.frame with columns for item1, item2 and gamma coefficient. Items can be variable.names, variable.columns or item numbers.}

\item{DIF}{A data.frame with columns for item, exogenous variable and gamma coefficient. Items and exogenous variables can be variable.names, variable.columns or item numbers.}
}
\value{
Returns a tbl_graph
}
\description{
Create a tidygraph object from a digram.object
}
\examples{
library(ggraph)
library(tidygraph)
dograph<-as_tbl_graph(DHP)
ggraph(dograph,layout="fr")+geom_edge_link()+geom_node_label(mapping = aes(label=label))
ggraph(dograph,layout="fr")+geom_edge_link()+geom_node_label(mapping = aes(label=name))

# Show arrows
ggraph(dograph,layout="fr")+geom_edge_link(end_cap = square(.5, 'cm'),arrow = arrow(angle=10,length=unit(.2,"cm")))+geom_node_label(mapping = aes(label=label))

# A digram.object with a testlet
dograph<-as_tbl_graph(code.testlet(DHP,"a b c"))
ggraph(dograph,layout="fr")+geom_edge_link(end_cap = square(.5, 'cm'),arrow = arrow(angle=10,length=unit(.2,"cm")))+geom_node_label(mapping = aes(label=label))

# Local dependecy and DIF
dograph<-as_tbl_graph(DHP,LD=data.frame(item1=c(5,3),item2=c(6,5),gamma=c(.53,-.38)),DIF=data.frame(item=c(4),exo=c(8),gamma=c(.27)))
ggraph(dograph,layout="fr")+geom_edge_link(mapping=aes(label=ifelse(!is.na(gamma),abs(gamma),""),alpha=ifelse(!is.na(gamma),gamma,1),color=ifelse(!is.na(gamma),2,1)),angle_calc="along",label_dodge=unit(.25,"cm"),end_cap = square(.5, 'cm'),arrow = arrow(angle=10,length=unit(.2,"cm")))+geom_node_label(mapping = aes(label=label))
}
\seealso{
\code{\link[tidygraph:as_tbl_graph]{tidygraph::as_tbl_graph()}}
}
