% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.object.R
\name{code.DIF}
\alias{code.DIF}
\title{Code items as having DIF}
\usage{
code.DIF(do, DIF.var, DIF.exo, append = F)
}
\arguments{
\item{do}{A digram.object}

\item{DIF.var}{String. The variables having DIF. A comma separated list of variable numbers, labels or names.}

\item{DIF.exo}{String. The variables causing DIF. A comma separated list of exogenous variable numbers, labels or names.}

\item{append}{Logical.}
}
\value{

}
\description{
Code items as having DIF
}
\details{
If more variables and exogenous variables are given, all possible combinations of these are set to exibit DIF.
}
\examples{
data(DHP)
do<-code.DIF(DHP,"a,b","under60")
}
