% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.object.R
\name{code.split}
\alias{code.split}
\title{Code items to be split (as having DIF)}
\usage{
code.split(do, split.var, split.on, append = F)
}
\arguments{
\item{do}{A digram.object}

\item{split.var}{String. The variables to split (having DIF). A comma separated list of variable numbers, labels or names.}

\item{split.on}{String. The exogenous variables to split on (causing DIF). A comma separated list of exogenous variable numbers, labels or names.}

\item{append}{Logical.}
}
\value{

}
\description{
Code items to be split (as having DIF)
}
\details{
If more variables and exogenous variables are given, all possible combinations of these are split.
}
\examples{
data(DHP)
do<-code.split(DHP,"a,b","under60")
}
