% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.fit.pattern.R
\name{delta.item}
\alias{delta.item}
\title{delta calculated from beta of the PCM}
\usage{
delta.item(x, single.item.params)
}
\arguments{
\item{x}{response (score on the item)}

\item{single.item.params}{a vector of item parameters (thresholds) (using the PCM parametrisation)}
}
\value{
Returns delta
}
\description{
delta calculated from beta of the PCM
}
\examples{
delta.igem(x,single.item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
