% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.fit.pattern.R
\name{delta.pattern}
\alias{delta.pattern}
\title{deltas calculated from betas of the PCM of items in a given response pattern}
\usage{
delta.pattern(xs, item.params)
}
\arguments{
\item{xs}{the response pattern (x1 ... xk)}

\item{item.params}{a matrix of item parameters (using the PCM parametrisation). Items in rows, threshold values in columns}
}
\value{
Returns the deltas in the power series parametrization of the Rasch Model from betas from the Partial Credit Model.
}
\description{
deltas calculated from betas of the PCM of items in a given response pattern
}
\examples{
delta.pattern(xs,item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
