% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proficiency.scale.R
\name{describe.items}
\alias{describe.items}
\title{Describe items}
\usage{
describe.items(
  item.descriptions = NULL,
  mod = NULL,
  item.params = NULL,
  outdir = getwd(),
  file.name = "item.descriptions"
)
}
\arguments{
\item{item.descriptions}{A matrix of item descriptions to edit.}

\item{mod}{An optional object of type tam.mml or eRm from which to extract the item.params.}

\item{item.params}{An optional matrix of item parameters to be used as a basis for the item.descriptions matrix.}

\item{outdir}{Where to save the output file.}

\item{file.name}{Filename of the output csv-file.}
}
\value{

}
\description{
Describe items
}
\examples{
#' mod<-digram.estimate(DHP)
describe.items(mod)
}
