% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.estimate.R
\name{digram.estimate}
\alias{digram.estimate}
\title{Estimate RDigram object using TAM}
\usage{
digram.estimate(
  do,
  items = NULL,
  groups = NULL,
  ncases = 0,
  constraint = "cases",
  use.package = c("TAM", "eRm"),
  collapse.testlets = F,
  init.model = NULL,
  tam.control = list(),
  sum0 = T,
  verbose = T,
  ...
)
}
\arguments{
\item{do}{A digram.object}

\item{items}{The items to include in the analysis}

\item{groups}{Names or column numbers of exogenous variables to use for grouping in TAM. If more names are given, all combinations of values are calculated and used as grouping variables.}

\item{ncases}{Number of cases to sample for the estimation (0 uses all cases)}

\item{constraint}{Constraint on "cases" or "items"}

\item{use.package}{Which R package to use for the estimation. TAM and eRm are implemented.}

\item{collapse.testlets}{Testlets are estimated using a bifactorial model in TAM and a data matrix in eRm. Setting collapse.testlets to TRUE calculates super-items instead and estimate a normal polytomous model.}

\item{init.model}{In TAM, the model that was output from an earlier estimation can be used to set sensible init-values for the estimation.}

\item{tam.control}{Use this to set control parameters in TAM estimation.}

\item{sum0}{Set to TRUE if you want eRm to sum the parameters to 0. If FALSE the first parameter is set to 0.}

\item{verbose}{Set to TRUE to get information about the estimation progress.}
}
\value{
Returns a TAM result object
}
\description{
Estimate RDigram object using TAM
}
\details{
Uses either the package TAM or eRm to estimate the model.
If items have been coded as testlets, a bifactorial model is used in TAM (\code{\link[=tam.fa]{tam.fa()}}). Otherwise \code{\link[=tam.mml]{tam.mml()}} is used for estimation.
In eRm, testlets are managed by creating an interaction parameter between the testlet items. In this case \code{\link[=LPCM]{LPCM()}} is used for estimation. This is also the case, if groups are provided. Otherwise \code{\link[=PCM]{PCM()}} is used for estimation.
}
\examples{
data(DHP)
do<-DHP
mod1<-digram.estimate(do)
summary(mod1)
do2<-code.LD(do,"ef")
mod2<-digram.estimate(do2)
summary(mod2)
mod1$deviance
mod2$deviance
mod1$deviance-mod2$deviance
}
\references{
Wang, W.-C., & Wilson, M. (2005). The Rasch Testlet Model. \emph{Applied Psychological Measurement}, 29(2), 126–149. https://doi.org/10.1177/0146621604271053
Rijmen, F. (2009). \emph{Three multidimensional models for testlet-based tests: Formal relations and an empirical comparison}. ETS Research Report Series, 2009(2), i–13. https://doi.org/10.1002/j.2333-8504.2009.tb02194.x
}
\seealso{
\code{\link[=tam.mml]{tam.mml()}}, \code{\link[=tam.fa]{tam.fa()}}, \code{\link[=PCM]{PCM()}}, \code{\link[=LPCM]{LPCM()}}
}
