% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.object.R
\name{digram.object}
\alias{digram.object}
\title{Create DIGRAM Object}
\usage{
digram.object(project=NULL,data=data.frame(),variables=colnames(data),filter.conditions=list(),recursive.structure=c(),comments="")
}
\arguments{
\item{project}{The name of the DIGRAM project}

\item{data}{The data.frame or matrix with the data}

\item{variables}{A vector of column names or numbers from the dataset to include in the recoded dataset (the order matters in the recoded data) \emph{or} a list of variables each element in the form list(variable.name="",variable.label="",ncat=0,category.names=c(),variable.type=c("nominal","ordinal"),minimum=0,maximum=1,cutpoints=c())}

\item{filter.conditions}{A list where names are variable numbers and values are \code{c(max,min)}. Only cases, for which all values of filter variables belong to the intervals defined by the corresponding minimum and maximum values (both included), will be used in the analysis.}

\item{recursive.structure}{A vector of cutpoints to define the recursive blocks}

\item{comments}{A string}
}
\value{
Returns a digram.object
}
\description{
Create a digram.object.
}
\examples{
library(iarm)
do<-digram.object(project = "desc2",data = desc2,variables = c(5:14,2:4,1),recursive.structure = c(10,13))
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
