% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.object.R
\name{digram.objects.combine}
\alias{digram.objects.combine}
\title{Combine several digram.objects into a combined.digram.objects.}
\usage{
digram.objects.combine(..., project = "")
}
\arguments{
\item{...}{digram.objects to combine}

\item{project}{Name of project}
}
\value{
Returns a combined.digram.objects
}
\description{
Combine several digram.objects into a combined.digram.objects.
}
\details{
This function is useful when preparing a multidimensional analysis. Each digram.object represents a dimension.¸
}
\examples{
do1<-DHP
do2<-DHP
digram.objects.combine(do1,do2)
}
