% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.person.p.R
\name{digram.person.p}
\alias{digram.person.p}
\title{Calculate probability of correct answer to item categories}
\usage{
digram.person.p(mod, item.params = NULL, thetas = NULL, ns = NULL)
}
\arguments{
\item{mod}{A TAM or eRm model}

\item{item.params}{Item params in PCM parametrization in a matrix or data.frame with categories in columns and persons in rows. If left out, item.params is taken from mod.}

\item{thetas}{A vector of person parameters. If left out, thetas is calculated from the model. It takes time on large datasets.}

\item{ns}{A vecor of the same length as thetas. Number of persons having each theta. If left out, it is counted from mod or set to 1.}
}
\value{
Returns a list of unique thetas, consisting of Theta, Npers (number of persons having this theta) and a data.frame of probabilities of getting points in the category.
}
\description{
Calculate probability of correct answer to item categories
}
\examples{
mod<-digram.estimate(DHP)
p<-digram.person.p(mod)
}
