% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{draw.plausible.response}
\alias{draw.plausible.response}
\title{Draw a plausible response based on item parameters}
\usage{
draw.plausible.response(
  item.params,
  param.type = c("pcm", "log.item.score", "multiplicative", "xsi"),
  R = 0
)
}
\arguments{
\item{item.params}{a matrix of item parameters (items in rows, thresholds in columns)}

\item{param.type}{Type of item parameters. One of pcm (RUMM2030), log.item.score (?), multiplicative (DIGRAM or RDigram, xsi (Conquest or TAM))}

\item{R}{Total score of the response}
}
\value{
Returns a plausible response pattern
}
\description{
Draw a plausible response based on item parameters
}
\examples{
item.params<-matrix(c(1,.5,1,1,1,2,1,4),nrow=4)
draw.plausible.response(item.params=item.params,param.type="multiplicative",R=2)
}
