% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJH5.R
\docType{data}
\name{EJH5}
\alias{EJH5}
\title{Erik Jørgen Hansen (EJH5)}
\format{
A data frame with 3151 rows and 18 variables.
\describe{
The included project variables are
D Income
C SRH
A ChronDis
B Unempl
F VocEduc
G School
I Intellig
J Urbaniza
K FamSES
L FamEduc
M Sex
}
}
\source{
The \code{EJH5} data were obtained from
Kreiner S. (2014) A guided tour through DIGRAM 3.11. Analysis of contingency tables by chain graph models.
}
\usage{
data(EJH5)
}
\value{
EJH5 is an object of class digram.object.
}
\description{
The project contains information from a panel study of living conditions in Denmark, containing information collected from adolescents in the eighth grade of the Danish public school in 1967, information on education later in life and finally information on health, income and unemployment collected in 1992.
}
\keyword{datasets}
