% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.fit.pattern.R
\name{gamma.pattern}
\alias{gamma.pattern}
\title{Gamma of a given pattern}
\usage{
\method{gamma}{pattern}(xs, item.params)
}
\arguments{
\item{xs}{A response pattern (scores on a series of items)}

\item{item.params}{a matrix of item parameters (using the PCM parametrisation). Items in rows, threshold values in columns}
}
\value{
Returns gamma.
}
\description{
Gamma of a given pattern
}
\examples{
gamma.pattern(xs,item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
