% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_matrix}
\alias{gamma_matrix}
\title{Gammas of item.parameters given total score using matrix calculations}
\usage{
gamma_matrix(
  item.params = NULL,
  param.type = c("multiplicative", "pcm", "log.item.score", "xsi"),
  R = NULL
)
}
\arguments{
\item{item.params}{a matrix of item parameters (using the PCM parametrisation). Items in rows, threshold values in columns}

\item{param.type}{Type of item parameters given. One of pcm (RUMM2030), log.item.score (?), multiplicative (DIGRAM or RDigram, xsi (Conquest or TAM))}

\item{R}{The total score for which to calculate the gamma parameter (set to NULL to get gammas for all possible total scores)}
}
\value{
Returns gamma for the given total score.
}
\description{
Gammas of item.parameters given total score using matrix calculations
}
\details{
Always use gamma.matrix instead of gamma.pattern. It is far more efficient.
}
\examples{
item.params<-matrix(c(1,.5,1,1,1,2,1,4),nrow=4)
gamma.matrix(item.params,"multiplicative",3)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
