% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{get.category.names}
\alias{get.category.names}
\title{Get category values and names}
\usage{
get.category.names(do, items = NULL, only.items = F, only.exos = F)
}
\arguments{
\item{do}{A digram.object}

\item{items}{The variable.numbers of the items or exogenous variables to provide the category names for}

\item{only.items}{Boolean. Only return items.}

\item{only.exos}{Boolean. Only return exos}
}
\value{
Returns the variable.names of the items/exogenous variables .
}
\description{
Get category values and names
}
\examples{
get.category.names(DHP,1:2)
}
