% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.item.params.R
\name{get.item.params}
\alias{get.item.params}
\title{Get item parameters and standard errors as a data.frame from a TAM or eRm object}
\usage{
get.item.params(obj)
}
\arguments{
\item{obj}{A TAM or eRm-object}
}
\value{
Returns a matrix of item parameters and standard errors, one row for each item
}
\description{
Get item parameters and standard errors as a data.frame from a TAM or eRm object
}
\examples{
data(DHP)
tamobj<-digram.estimate(DHP)
get.item.params(tamobj)
}
