% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{get.labels}
\alias{get.labels}
\title{Get variable labels}
\usage{
get.labels(do, items = NULL)
}
\arguments{
\item{do}{A digram.object}

\item{items}{The variable.numbers of the items or exogenous variables to provide the labels for}
}
\value{
Returns the variable.labels of the items/exogenous variables .
}
\description{
Get variable labels
}
\examples{
get.labels(DHP,1:2)
}
