% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.thresholds.R
\name{get.thresholds}
\alias{get.thresholds}
\title{Get xsi item parameters from TAM object}
\usage{
get.thresholds(tamobj)
}
\arguments{
\item{tamobj}{A TAM-object}
}
\value{
Returns a matrix of xsi parameters
}
\description{
Get xsi item parameters from TAM object
}
\examples{
data(DHP)
tamobj<-digram.estimate(DHP)
get.thresholds(tamobj)
}
