% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.R
\name{get.variable.names}
\alias{get.variable.names}
\title{Get variable names}
\usage{
get.variable.names(do, items = NULL)
}
\arguments{
\item{do}{A digram.object}

\item{items}{The variable.numbers of the items or exogeneous variables to provide the names for}
}
\value{
Returns the variable.names of the items/exogeneous variables .
}
\description{
Get variable names
}
\examples{
get.variable.names(DHP,1:2)
}
