% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativeproportionalfitting.R
\name{ipf}
\alias{ipf}
\title{Iterative Proportional Fitting (ipf)}
\usage{
ipf(
  X = matrix(),
  Xhat = matrix(0.5, nrow = nrow(X), ncol = ncol(X)),
  Delta = 1,
  func = function(i, j, A, B, Delta) {
     1
 },
  anchor = NULL,
  v = 1
)
}
\arguments{
\item{X}{A matrix, e.g. of persons responses (rows) to items (columns)}

\item{Xhat}{Initial values of the solution (optional). Defaults to 0.5 in all cells}

\item{Delta}{Interaction parameters. A matrix or scalar. Defaults to 1.}

\item{func}{A function of parameters i, j (row and column numbers), A, B (current estimations),  and Delta.}

\item{anchor}{A function to anchor A or B estimates. Defaults to NULL.}

\item{v}{Verbose level (0: none, 1: progression, 2: verbose)}
}
\value{
Returns a list of estimates Ahat, Bhat, and Xhat
}
\description{
Iterative Proportional Fitting (ipf)
}
\details{
This function does the estimation. Used by specific estimation functions.
}
\examples{
X<-matrix(data=c(0,0,0,1,1,1,0,1,0,0,0,1,
}
\references{
Svend Kreiner (manuscript, 2023). \emph{Appendix H - Estimation of Rasch parameters by iterative proportional fitting}.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
