% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.correlations.R
\name{item.correlations}
\alias{item.correlations}
\title{Item correlations}
\usage{
item.correlations(do=NULL,resp=NULL,items=1:do$recursive.structure[1],exo=(do$recursive.structure[1]+1):do$recursive.structure[2])
}
\arguments{
\item{do}{an object of class \code{digram.object}}

\item{resp}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include as items in the analysis \emph{or} a character vector of variable labels}

\item{exo}{A vector of columns from the recoded data to include as exogenous variables in the analysis \emph{or} a character vector of variable labels}

\item{max.name.length}{Maximum length of item names (to be printed in tables)}

\item{accept.na}{A boolean. Include cases with missing values in responses}

\item{verbose}{Print results}
}
\value{
Returns a list of correlations
}
\description{
Calculate item correlations, item-rest correlations and correlations between items and exogenous variables
}
\details{
First step in item screening: Analysis of consistency (Positive correlations)
\describe{
\item{M1}{\eqn{Y_i} and \eqn{Y_j} are positively correlated for all pairs of items}
\item{M2}{\eqn{Y_a} is positively monotonically related to the rest-score \eqn{R_a} and all subscores \eqn{S_B} where \eqn{Y_a} \eqn{\not\in}{is not an element in} B}
\item{M3}{If \eqn{X} is positively related to \eqn{\theta}, then \eqn{X} will also be positively related to \eqn{S}, to all subscores, \eqn{S_A}, and all item responses \eqn{Y_i}}
}
}
\examples{
library(iarm)
do<-digram.object(project = "desc2",data = desc2,variables = c(5:14,2:4,1),recursive.structure = c(10,13))
item.correlations(do)
}
\references{
Kreiner, S. & Christensen, K.B. (2011). Item Screening in Graphical Loglinear Rasch Models. \emph{Psychometrika}, vol. 76, no. 2, pp. 228-256. DOI: 10.1007/s11336-9203-Y
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
