% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proficiency.scale.R
\name{item.descriptions.from.params}
\alias{item.descriptions.from.params}
\title{Produce a template matrix of item descriptions from item parameters}
\usage{
item.descriptions.from.params(item.params = NULL, mod = NULL)
}
\arguments{
\item{item.params}{A matrix of item parameters, items in rows and categories in columns. Item-names as rownames.}

\item{mod}{An optional object of type tam.mml or eRm from which to extract the item.params.}
}
\value{
Returns a matrix of item descriptions.
}
\description{
Produce a template matrix of item descriptions from item parameters
}
\examples{
mod<-digram.estimate(DHP)
item.descriptions.from.params(mod=mod)
}
