% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.item.parameters.R
\name{item.param.convert}
\alias{item.param.convert}
\title{Convert item parameters from one parametrization to another}
\usage{
item.param.convert(
  item.params = c(),
  from = c("pcm", "log.item.score", "multiplicative", "xsi"),
  to = c("pcm", "log.item.score", "multiplicative", "xsi")
)
}
\arguments{
\item{item.params}{a matrix of item parameters (items in rows, thresholds in columns)}

\item{from, to}{Type of item parameters. One of pcm (RUMM2030), log.item.score (?), multiplicative (DIGRAM or RDigram, xsi (Conquest or TAM))}
}
\value{
Returns item parameters in the parametrization specified in \emph{to}
}
\description{
Convert item parameters from one parametrization to another
}
\examples{
item.params<-matrix(c(0,1,2,3,1,2,3,4),nrow=4)
item.param.convert(item.params=item.params,from="xsi",to="multiplicative")
}
