% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.dependency.R
\name{local.dependency}
\alias{local.dependency}
\alias{Additional}
\alias{aliases}
\alias{through}
\alias{which}
\alias{users}
\alias{can}
\alias{find}
\alias{documentation}
\title{Detect Local dependency}
\arguments{
\item{do}{an object of class \code{digram.object}}
}
\value{
list of local dependencies
}
\description{
One-line function description
}
\details{
Multiline function description (indent after the first line)
}
\note{
Any notes about the operation of the function
}
\examples{
Examples using the function
}
\references{
References concerning the methodology employed by function
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
