% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.independency.R
\name{local.independency}
\alias{local.independency}
\title{Detect local dependency}
\usage{
local.independency(do=NULL,recoded=NULL,items=1:do$recursive.structure[1],exo=(do$recursive.structure[1]+1):do$recursive.structure[2])
}
\arguments{
\item{do}{an object of class \code{digram.object}}

\item{recoded}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include as items in the analysis \emph{or} a character vector of variable labels}

\item{exo}{A vector of columns from the recoded data to include as exogenous variables in the analysis \emph{or} a character vector of variable labels}

\item{digits}{Number of digits in table}

\item{verbose}{Print results}
}
\value{
Returns a list of local dependencies
}
\description{
Investigate items for local independence
}
\details{
Second step in item screening: Analysis of DIF and local dependency
\describe{
\item{C2}{Y i ⊥X j |S for all i = 1 . . . k and j = 1 ...m}
\item{C4}{Y a ⊥Y b |R a and Y a ⊥Y b |R b}
}
}
\examples{
local.independency(DHP)
}
\references{
Kreiner, S. & Christensen, K.B. (2011). Item Screening in Graphical Loglinear Rasch Models. \emph{Psychometrika}, vol. 76, no. 2, pp. 228-256. DOI: 10.1007/s11336-9203-Y
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
