% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.fit.pattern.R
\name{person.fit.pattern}
\alias{person.fit.pattern}
\title{Person fit based on response pattern}
\usage{
person.fit.pattern(
  do = NULL,
  resp = NULL,
  items = NULL,
  item.params = matrix()
)
}
\arguments{
\item{do}{a DIGRAM object}

\item{resp}{a matix of responses (if no DIGRAM object is supplied)}

\item{items}{a vector of items to use}

\item{item.params}{a matrix of item parameters (using the PCM parametrisation). Items in rows, threshold values in columns}
}
\value{
Returns a list of results for each respondent, consisting of response pattern probability and the sum of probabilities of patterns with lower probability (montecarlo probability)
}
\description{
Person fit based on response pattern
}
\examples{
person.fit.pattern(do=do, item.params=item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
