% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf3.R
\docType{data}
\name{pf3}
\alias{pf3}
\title{Physical functioning subscale}
\format{A data frame with 1049 rows and 25 variables of which 12 are used.
The scale summarizes responses to the following ten items:
\describe{
Does your health now limit you in these activities? If so, how much?
<<<<<<< HEAD
\item{A}{PF1: Vigorous activities}
\item{B}{PF2: Moderate activities}
\item{C}{PF3: Lifting or carrying groceries}
\item{D}{PF4: Climbing several flights of stairs}
\item{E}{PF5: Climbing one flight of stairs}
\item{F}{PF6: Bending, kneeling, or stooping}
\item{G}{PF7: Walking more than a mile}
\item{H}{PF8: Walking several blocks}
\item{I}{PF9: Walking one block}
\item{J}{PF10: Bathing or dressing yourself}
The responses to these questions were coded in the following way:
\item{0}{Limited a lot}
\item{1}{Limited a little}
\item{2}{Not limited}
=======
\item{A) PF1: Vigorous activities}
\item{B) PF2: Moderate activities}
\item{C) PF3: Lifting or carrying groceries}
\item{D) PF4: Climbing several flights of stairs}
\item{E) PF5: Climbing one flight of stairs}
\item{F) PF6: Bending, kneeling, or stooping}
\item{G) PF7: Walking more than a mile}
\item{H) PF8: Walking several blocks}
\item{I) PF9: Walking one block}
\item{J) PF10: Bathing or dressing yourself}
The responses to these questions were coded in the following way:
\item{0: Limited a lot}
\item{1: Limited a little}
\item{2: Not limited}
>>>>>>> 828497a089c4831be5f2b6b3c5e41b0e2862a181
so that a low score indicates physical impairment. Gender and Age are also included in this project.
}}
\source{
The \code{pf3} data were obtained from
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\usage{
data(pf3)
}
\value{
pf3 is an object of class digram.object.
}
\description{
The dataset originated in a Danish Health survey. This is the SF36 subscale measuring physical functioning.
}
\keyword{datasets}
