% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirlsDK_2016_BK16.R
\docType{data}
\name{pirlsDK_2016_BK16}
\alias{pirlsDK_2016_BK16}
\title{PIRLS data for Denmark 2016}
\format{
A data frame with 557 rows and 34 variables. 32 response variables, 2 context variables.
The data was collected in IEA's PIRLS 2016 literacy tests, released items. Information about the test and the reading material: https://dpu.au.dk/forskning/internationaleundersoegelser/pirls/pirls-2016/materialer
\describe{
The booklet can be downloaded here: https://dpu.au.dk/fileadmin/edu/Forskning/Internationale_undersoegelser/PIRLS/PIRLS_2016/PIRLS_2016__Laesebog__anvendt_i_4._klasse.pdf
}
}
\source{
https://timssandpirls.bc.edu/pirls2016/index.html
}
\usage{
data(pirlsDK_2016_BK16)
}
\value{
pirlsDK_2016_BK16 is an object of class digram.object.
}
\description{
PIRLS 2016 literacy test, released items, Booklet 16, for Danish students.
}
\keyword{datasets}
