% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.digram.object.R
\name{print.digram.object}
\alias{print.digram.object}
\title{Print DIGRAM object}
\usage{
\method{print}{digram.object}(do = NULL)
}
\arguments{
\item{do}{a DIGRAM object}
}
\value{
Returns nothing
}
\description{
Prints information about a DIGRAM object
}
\details{
Prints
}
\note{
Any notes about the operation of the function
}
\examples{
print(DHP)
}
\references{
References concerning the methodology employed by function
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
