% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDigram.R
\docType{package}
\name{RDigram-package}
\alias{RDigram-package}
\alias{RDigram}
\title{Functions for item analysis}
\description{
RDigram provides tools for analysis of items from achievement tests and surveys.
RDigram is an implementation of a subset of the functions in the DIGRAM statistical
program by Svend Kreiner.
}
