% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.digram.R
\name{read.digram}
\alias{read.digram}
\title{Read DIGRAM Object}
\usage{
read.digram(project=NULL,path="")
}
\arguments{
\item{project}{The name of the DIGRAM project from which to load data}

\item{path}{Path to the project files}
}
\value{
Returns a digram.object
}
\description{
Reads data from DIGRAM files and create a digram.object.
}
\examples{
do<-read.digram("DHP",path = "DHP")
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
