% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proficiency.scale.R
\name{read.item.descriptions}
\alias{read.item.descriptions}
\title{Read item descriptions from file}
\usage{
read.item.descriptions(mod, file.name = "item.descriptions.csv")
}
\arguments{
\item{mod}{A TAM model to be used when filtering the item descriptions (see details)}

\item{file.name}{An optional filename of the item descriptions. Default is 'item.descriptions.csv'.}
}
\value{
Returns item descriptions from file.
}
\description{
Read item descriptions from file
}
\details{
Item descriptions are filtered so descriptions of unused items are left out, and non-existing descriptions are given the item-name and cathegory
}
\examples{
mod<-digram.estimate(DHP)
read.item.descriptions(mod=mod)
}
