% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.information.R
\name{score.information}
\alias{score.information}
\title{Print score information}
\usage{
score.information(do=NULL,resp=NULL,items=1:length(do$variables),accept.na = F)
}
\arguments{
\item{do}{An object of class \code{digram.object}}

\item{resp}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include in the analysis \emph{or} a character vector of variable labels}

\item{do.testlets}{Bolean. If TRUE, testlets are combined to superitems.}

\item{do.split}{Bolean. If TRUE, items coded as split are split.}

\item{accept.na}{A boolean. Include cases with missing values in responses}

\item{rownamewidth}{Width of item names column}
}
\value{
Returns an object of items.scores,scoredist,meansum,score.distribution, and score.groups.
Prints Average item scores and score distribution and Score groups for tests of Rasch models
}
\description{
Prints information about the recoded data.
}
\examples{
score.information(do,items="abcdef")
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
