% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.digram.object.R
\name{summary.digram.object}
\alias{summary.digram.object}
\title{Summarize DIGRAM object}
\usage{
\method{summary}{digram.object}(do = NULL)
}
\arguments{
\item{do}{a DIGRAM object}
}
\value{
Returns nothing
}
\description{
Summarizes information about a DIGRAM object
}
\examples{
summary(DHP)
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
