% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMSSTaiAus.R
\docType{data}
\name{TIMSSTaiAus}
\alias{TIMSSTaiAus}
\title{TIMSS data for Taiwan and Austrailia}
\format{
A data frame with 1773 rows and 14 variables. 11 response variables, 3 context variables.
The data was collected in IEA's TIMSS 2011 mathematics tests, released items, blocks 1 and 14, for Taiwan and Australian students.
\describe{
The booklet can be downloaded here: https://www.edmeasurementsurveys.com/TAM/Tutorials/data/TIMSS2011G8Block1.pdf
}
}
\source{
https://timss.bc.edu/timss2011/index.html
}
\usage{
data(TIMSSTaiAus)
}
\value{
TIMSSTaiAus is an object of class digram.object.
}
\description{
TIMSS 2011 mathematics tests, released items, blocks 1 and 14, for Taiwan and Australian students. From TAM tutorial https://www.edmeasurementsurveys.com/TAM/Tutorials/5PartialCredit.htm. Coded according to the tutorial.
}
\keyword{datasets}
