% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.update.R
\name{variable.delete}
\alias{variable.delete}
\title{Delete variable}
\usage{
variable.delete(do = NULL, variable.to.delete = NULL)
}
\arguments{
\item{do}{A digram.object}

\item{variable.to.delete}{The number or name of the variable(s) which should be deleted}
}
\value{
Returns a DIGRAM object with (a) deleted variable(s)
}
\description{
Delete one or more variable(s) in a DIGRAM Object.
}
\details{
A set of variables can be deleted by providing a vector of variable names/numbers.
}
\examples{
data(DHP)
DHP<-variable.delete(do=DHP,variable.to.delete="dhp36")

}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
