% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.update.R
\name{variable.update}
\alias{variable.update}
\title{Update variable}
\usage{
variable.update(do=NULL,variable=NULL,variable.name=NULL,variable.label=NULL,category.names=c(),variable.type=c("ordinal","nominal"),minimum=NULL,maximum=NULL,cutpoints=c())
}
\arguments{
\item{do}{A digram.object}

\item{variable.to.update}{The number or name of the variable that should be updated}

\item{variable.name}{New name for the variable (string)}

\item{variable.label}{New label for the variable (one or more uppercase letters)}

\item{variable.type}{Type of variable ("ordinal" or "nominal")}

\item{minimum}{Smallest value of the variable (lower values are coded NA)}

\item{maximum}{Largest value of the variable (lower values are coded NA)}

\item{cutpoints}{Cutpoints used to recode the variable
Category 1: minimum <= values <= cutpoint(1)
Category 2: cutpoint(1) < values <= cutpoint(2)
...
Category N: cutpoint(n) < values <= maximum}
}
\value{
Returns a DIGRAM object with an updated variable
}
\description{
Update variable in DIGRAM Object.
}
\examples{
DHP<-variable.update(do=DHP,variable.to.update="dhp36",cutpoints=c(2,3))
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
