% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.digram.R
\name{write.digram}
\alias{write.digram}
\title{Write DIGRAM Object}
\usage{
write.digram(do = NULL, path = "", filename = do$project)
}
\arguments{
\item{do}{The digram.object}

\item{path}{Path to where to save the files}

\item{filename}{Optional. Set to project name if not set.}
}
\value{
Returns TRUE on success
}
\description{
Writes a RDigram object to DIGRAM files.
}
\examples{
do<-write.digram(do,path = "DHP")
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
