% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.item.parameters.R
\name{xsi.to.multiplicative}
\alias{xsi.to.multiplicative}
\title{Multiplicative item parameters calculated from xsi parameters (so-called Conquest parametrization)}
\usage{
xsi.to.multiplicative(item.params)
}
\arguments{
\item{item.params}{a matrix of xsi item parameters (using the Conquest parametrisation). Items in rows, threshold values in columns}
}
\value{
Returns the deltas in the power series parametrization (e.g. from DIGRAM) of the Rasch Model from the xsi parameters (e.g. from Conquest and TAM).
}
\description{
Multiplicative item parameters calculated from xsi parameters (so-called Conquest parametrization)
}
\examples{
item.params<-matrix(c(1,.5,1,1,1,2,1,4),nrow=4)
xsi.to.multiplicative(item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
