% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.item.parameters.R
\name{delta.item}
\alias{delta.item}
\title{Delta of the multiplicative parametrization calculated from xsi parameter(Conquest parametrization)}
\usage{
delta.item(single.item.params)
}
\arguments{
\item{single.item.params}{a vector of item parameters (thresholds) (using the Conquest parametrisation)}
}
\value{
Returns delta
}
\description{
Delta of the multiplicative parametrization calculated from xsi parameter(Conquest parametrization)
}
\examples{
single.item.params<-c(1,1,0.5)
delta.item(single.item.params=single.item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
