% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.estimate.R
\name{digram.estimate}
\alias{digram.estimate}
\title{Estimate RDigram object using TAM}
\usage{
digram.estimate(do, items = NULL)
}
\arguments{
\item{do}{A digram.object}

\item{items}{The items to include in the analysis}
}
\value{
Returns a TAM result object
}
\description{
Estimate RDigram object using TAM
}
\examples{
data(DHP)
do<-DHP
mod1<-digram.estimate(do)
summary(mod1)
do2<-code.LD(do,"ef")
mod2<-digram.estimate(do2)
summary(mod2)
mod1$deviance
mod2$deviance
mod1$deviance-mod2$deviance
}
\references{
Wang, W.-C., & Wilson, M. (2005). The Rasch Testlet Model. \emph{Applied Psychological Measurement}, 29(2), 126–149. https://doi.org/10.1177/0146621604271053
Rijmen, F. (2009). \emph{Three multidimensional models for testlet-based tests: Formal relations and an empirical comparison}. ETS Research Report Series, 2009(2), i–13. https://doi.org/10.1002/j.2333-8504.2009.tb02194.x
}
\seealso{
tam.mml(), tam.fa()
}
