% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digram.wrightmap.R
\name{digram.wrightmap}
\alias{digram.wrightmap}
\title{Draw a Wrightmap}
\usage{
digram.wrightmap(
  mod,
  do = NULL,
  dimname = NULL,
  cols = collist,
  colscheme = "(.*)_.*",
  verbose = T
)
}
\arguments{
\item{mod}{Output from tam.mml or tam.fa.}

\item{do}{A digram.object. Used to create title. Alternative to dimname}

\item{dimname}{A string. Used to create title. Alternative to do.}

\item{cols}{A vector of colors.}

\item{colscheme}{A regular expression as string. Used to give items of the same type, same colors.}

\item{verbose}{Boolean. Set to TRUE to get information about progress.}
}
\value{
Returns a list of item difficulties as thresholds and person abilities as thetas
}
\description{
Draw a Wrightmap
}
\examples{
mod<-digram.estimate(DHP)
digram.wrightmap(mod=mod,do=DHP)
}
