% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.update.R
\name{get.column.no}
\alias{get.column.no}
\title{Get column number of a variable in the recoded data}
\usage{
get.column.no(do, variable.name.num)
}
\arguments{
\item{do}{A DIGRAM object}

\item{variable.name.num}{The number, label or name of the variable to search for}
}
\value{
Returns the column.number
}
\description{
Get column number of a variable in the recoded data
}
\examples{
data(DHP)
get.column.no(DHP,"dhp36")
}
