% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.correlations.R
\name{item.correlations}
\alias{item.correlations}
\title{Item correlations}
\usage{
item.correlations(do=NULL,recoded=NULL,items=1:do$recursive.structure[1],exo=(do$recursive.structure[1]+1):do$recursive.structure[2])
}
\arguments{
\item{do}{an object of class \code{digram.object}}

\item{recoded}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include as items in the analysis \emph{or} a character vector of variable labels}

\item{exo}{A vector of columns from the recoded data to include as exogenous variables in the analysis \emph{or} a character vector of variable labels}

\item{accept.na}{A boolean. Include cases with missing values in responses}

\item{verbose}{Print results}
}
\value{
Returns a list of correlations
}
\description{
Calculate item correlations, item-rest correlations and correlations between items and exogenous variables
}
\details{
First step in item screening: Analysis of consistency (Positive correlations)
\describe{
\item{M1}{Yi and Yj are positively correlated for all pairs of items}
\item{M2}{Ya is positively monotonically related to the rest-score Ra and all subscores SB where Ya is not an element in B}
\item{M3}{If X is positively related to Θ, then X will also be positively related to S, to all subscores, SA , and all item responses Yi}
}
}
\examples{
library(iarm)
do<-digram.object(project = "desc2",data = desc2,variables = c(5:14,2:4,1),recursive.structure = c(10,13))
item.correlations(do)
}
\references{
Kreiner, S. & Christensen, K.B. (2011). Item Screening in Graphical Loglinear Rasch Models. \emph{Psychometrika}, vol. 76, no. 2, pp. 228-256. DOI: 10.1007/s11336-9203-Y
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
