% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.R
\name{item.DIF}
\alias{item.DIF}
\title{Detect Differential Item Functioning}
\usage{
item.DIF(do=NULL,resp=NULL,items=1:do$recursive.structure[1],exo=(do$recursive.structure[1]+1):do$recursive.structure[2],verbose=T)
}
\arguments{
\item{do}{an object of class \code{digram.object}}

\item{resp}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include as items in the analysis \emph{or} a character vector of variable labels}

\item{exo}{A vector of columns from the recoded data to include as exogenous variables in the analysis \emph{or} a character vector of variable labels}

\item{p.adj}{the kind of multiple p-value testing adjustment to be used (one of "BH","holm", "hochberg", "hommel", "bonferroni", "BY", "none").}

\item{max.name.length}{Maximum length of item names (to be printed in tables)}

\item{digits}{Number of digits in table}

\item{only.significant}{Only list fit values significantly different from 1}

\item{verbose}{Print results}
}
\value{
Returns a list of DIF-information
}
\description{
Detect Differential Item Functioning (DIF)
}
\details{
Second step in item screening: Analysis of DIF and local dependency
\describe{
\item{C2}{\eqn{Y_i \perp X_j \divides S}{Y_i ⊥ X_j | S} for all \eqn{i = 1} \eqn{\ldots}{...} \eqn{k} and \eqn{j = 1} \eqn{\ldots}{...} \eqn{m}}
\item{C4}{\eqn{Y_a \perp Y_b \divides R_a}{Y_a ⊥ Y_b |R_a} and \eqn{Y_a \perp Y_b \divides R_b}{Y_a ⊥ Y_b | R_b}}
Conditional independence of A and B given C is denoted as \eqn{A \perp B \divides C}{A ⊥ B | C}.
}
Use local.independence() to detect local dependency
}
\examples{
item.DIF(DHP)
}
\references{
Kreiner, S. & Christensen, K.B. (2011). Item Screening in Graphical Loglinear Rasch Models. \emph{Psychometrika}, vol. 76, no. 2, pp. 228-256. DOI: 10.1007/s11336-9203-Y
}
\seealso{
\code{\link{partgam_DIF}}
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
