% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.independence.R
\name{local.independence}
\alias{local.independence}
\title{Detect local dependence}
\usage{
local.independence(do=NULL,resp=NULL,items=1:do$recursive.structure[1],digits=2,verbose=T)
}
\arguments{
\item{do}{an object of class \code{digram.object}}

\item{resp}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include as items in the analysis \emph{or} a character vector of variable labels}

\item{p.adj}{the kind of multiple p-value testing adjustment to be used (one of "BH","holm", "hochberg", "hommel", "bonferroni", "BY", "none").}

\item{digits}{Number of digits in table}

\item{only.significant}{Only list fit values significantly different from 1}

\item{verbose}{Print results}
}
\value{
Returns a list of local dependencies
}
\description{
Investigate items for local independence
}
\details{
Second step in item screening: Analysis of DIF and local dependence
\describe{
\item{C2}{Y i ⊥X j |S for all i = 1 . . . k and j = 1 ...m}
\item{C4}{Y a ⊥Y b |R a and Y a ⊥Y b |R b}
}
#' Use item.DIF() for detection of Differential Item Functioning
}
\examples{
local.independence(DHP)
}
\references{
Kreiner, S. & Christensen, K.B. (2011). Item Screening in Graphical Loglinear Rasch Models. \emph{Psychometrika}, vol. 76, no. 2, pp. 228-256. DOI: 10.1007/s11336-9203-Y
}
\seealso{
\code{\link{partgam_LD}}
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
