% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.independence.R
\name{local.independence}
\alias{local.independence}
\title{Detect local dependence}
\usage{
local.independence(do=NULL,resp=NULL,items=1:do$recursive.structure[1],digits=2,verbose=T)
}
\arguments{
\item{do}{an object of class \code{digram.object}}

\item{resp}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include as items in the analysis \emph{or} a character vector of variable labels}

\item{p.adj}{the kind of multiple p-value testing adjustment to be used (one of "holm", "BH","hochberg", "hommel", "bonferroni", "BY", "none"), see \code{\link[=p.adjust]{p.adjust()}}.}

\item{digits}{Number of digits in table}

\item{max.name.length}{Maximum length of item names (to be printed in tables)}

\item{only.significant}{Only list fit values significantly different from 1}

\item{use.names}{Use item names instead of item labels as node labels}

\item{verbose}{Print results}

\item{extra.verbose}{Print warnings in PDF and HTML-output}

\item{saved.result}{To avoid repeated calculation, you can provide a saved version of the analysis (returned from local.independence())}

\item{caption}{Caption for the LD table (in Rmarkdown)}
}
\value{
Returns a list of local dependencies
}
\description{
Investigate items for local independence
}
\details{
Second step in item screening: Analysis of DIF and local dependence
\describe{
\item{C4}{\eqn{Y_a \bot Y_b \mid R_a}{Y_a ⊥ Y_b | R_a} and \eqn{Y_a \bot Y_b \mid R_b}{Y_a ⊥ Y_b | R_b}}
Conditional independence of A and B given C is denoted as \eqn{A \bot B \mid C}{A ⊥ B | C}.
}
Use \code{\link[=item.DIF]{item.DIF()}} for detection of Differential Item Functioning

If you want to use this function in R Markdown or Bookdown, you need to use xelatex as latex engine, and you need to force dev to use cairo_pdf or png.
Add this in you yaml header:

\code{output:\cr
  pdf_document: \cr
    latex_engine: xelatex}

Add this in your setup chunk:

\code{knitr::opts_chunk$set(echo = TRUE, dev = "cairo_pdf", dpi = 300)}
}
\examples{
local.independence(DHP)
}
\references{
Kreiner, S. & Christensen, K.B. (2011). Item Screening in Graphical Loglinear Rasch Models. \emph{Psychometrika}, vol. 76, no. 2, pp. 228-256. DOI: 10.1007/s11336-9203-Y
}
\seealso{
\code{\link[=partgam_LD]{partgam_LD()}}, \code{\link[=item.DIF]{item.DIF()}}
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
