% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.fit.pattern.R
\name{person.fit.pattern}
\alias{person.fit.pattern}
\title{Person fit based on response pattern}
\usage{
person.fit.pattern(
  do = NULL,
  resp = NULL,
  items = NULL,
  item.params = matrix(),
  param.type = c("pcm", "log.item.score", "multiplicative", "xsi"),
  num.montecarlo = 0,
  verbose = T
)
}
\arguments{
\item{do}{a DIGRAM object}

\item{resp}{a matix of responses (if no DIGRAM object is supplied)}

\item{items}{a vector of items to use}

\item{item.params}{a matrix of item parameters. Items in rows, thresholds in columns}

\item{param.type}{Type of item parameters given. One of pcm (RUMM2030), log.item.score (?), multiplicative (DIGRAM or RDigram, xsi (Conquest or TAM))}

\item{num.montecarlo}{the number of iterations if the calculation of alternative patterns should be done using a montecarlo solution. 0 to do all patterns.}

\item{verbose}{set to TRUE if you want to follow the progression}
}
\value{
Returns a list of results for each respondent, consisting of response pattern probability and the p-value of getting this pattern or a pattern of lower probability.
}
\description{
Person fit based on response pattern
}
\examples{
data(DHP)
person.fit.pattern(do=DHP, item.params=item.params)
}
\references{
Jeppe Bundsgaard & Svend Kreiner (2019). \emph{Undersøgelse af De Nationale Tests måleegenskaber}. 2nd Ed. Copenhagen: DPU, Aarhus University.
}
\author{
Jeppe Bundsgaard & Svend Kreiner
}
