% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.information.R
\name{score.information}
\alias{score.information}
\title{Print score information}
\usage{
<<<<<<< HEAD
score.information(do=NULL,recoded=NULL,items=1:length(do$variables),accept.na = F)
=======
score.information(do=NULL,recoded=NULL,items=1:length(do$variables))
>>>>>>> 828497a089c4831be5f2b6b3c5e41b0e2862a181
}
\arguments{
\item{do}{An object of class \code{digram.object}}

\item{recoded}{A data.frame or matrix of recoded data (only used if \code{do} is \code{NULL})}

\item{items}{A vector of columns from the recoded data to include in the analysis \emph{or} a character vector of variable labels}
<<<<<<< HEAD

\item{accept.na}{A boolean. Include cases with missing values in responses}
=======
>>>>>>> 828497a089c4831be5f2b6b3c5e41b0e2862a181
}
\value{
Returns NULL.
Prints Average item scores and score distribution and Score groups for tests of Rasch models
}
\description{
Prints information about the recoded data.
}
\examples{
score.information(do,items="abcdef")
}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
