% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.update.R
\name{variable.update}
\alias{variable.update}
\title{Update variable}
\usage{
variable.update(
  do = NULL,
  variable.to.update = NULL,
  variable.name = NULL,
  variable.label = NULL,
  category.names = NULL,
  variable.type = NULL,
  minimum = NULL,
  maximum = NULL,
  cutpoints = NULL
)
}
\arguments{
\item{do}{A digram.object}

\item{variable.to.update}{The number or name of the variable(s) which should be updated}

\item{variable.name}{New name for the variable (string)}

\item{variable.label}{New label for the variable (one or more uppercase letters)}

\item{variable.type}{Type of variable ("ordinal" or "nominal")}

\item{minimum}{Smallest value of the variable (lower values are coded NA)}

\item{maximum}{Largest value of the variable (lower values are coded NA)}

\item{cutpoints}{Cutpoints used to recode the variable
Category 1: minimum <= values <= cutpoint(1)
Category 2: cutpoint(1) < values <= cutpoint(2)
...
Category N: cutpoint(n) < values <= maximum}
}
\value{
Returns a DIGRAM object with (an) updated variable(s)
}
\description{
Update one or more variable(s) in a DIGRAM Object.
}
\details{
A set of variables can be updated by providing a vector of variable names/numbers to update and ordered lists with their new properties. For variable.type, cutpoints, minimum and maximum you can provide one property to give all variables.
}
\examples{
data(DHP)
DHP<-variable.update(do=DHP,variable.to.update="dhp36",cutpoints=c(2,3))
DHP<-variable.update(do=DHP,variable.to.update=c(1,3),variable.label=list("32","36"))

}
\references{
Kreiner, S. (2003). \emph{Introduction to DIGRAM}. Dept. of Biostatistics, University of Copenhagen.
}
\author{
Jeppe Bundsgaard \href{mailto:jebu@edu.au.dk}{jebu@edu.au.dk}
}
